Updated index.js and get_Color_Pair.js
